VIX-3190 Addressing review comments
VIX-3190 Addressing review comments
