+++ Modules/Effect/Liquid/Emitter/IEmitter.cs
@@ -60,30 +60,18 @@ namespace VixenModules.Effect.Liquid
     /// <summary>
     /// Maintains the properties and methods of an emitter.
     /// </summary>
-	public interface IEmitter: ICloneable
-	{
+	public interface IEmitter: ICloneable, IMarkCollectionExpandoObject
+	{		
 		/// <summary>
-      /// Parent effect of the emitter.
-      /// This property is used to register for mark events.
-      /// </summary>
-		BaseEffect Parent { get; set; }
-		
-      /// <summary>
-      /// Returns a clone of the emitter.
-      /// </summary>        
+		/// Returns a clone of the emitter.
+		/// </summary>        
 		IEmitter CreateInstanceForClone();
 
 		/// <summary>
 		/// Initializes the browseable state of all the emitter properties.
 		/// </summary>
 		void InitAllAttributes();
-
-		/// <summary>
-		/// Updates the selected mark name collection for the emitter.
-		/// This method handles when a IMarkCollection is renamed.
-		/// </summary>
-		void UpdateSelectedMarkCollectionName();
-
+		
 		/// <summary>
 		/// Type of particles emitted by the emitter.
 		/// </summary>
@@ -119,9 +107,9 @@ namespace VixenModules.Effect.Liquid
 		/// </summary>
 		Curve ParticleVelocity { get; set; }
         
-      /// <summary>
-      /// Whether the emitter is animated.
-      /// </summary>
+		/// <summary>
+		/// Whether the emitter is animated.
+		/// </summary>
 		bool Animate { get; set; }
 
 		/// <summary>
@@ -144,84 +132,84 @@ namespace VixenModules.Effect.Liquid
 		/// </summary>
 		EdgeHandling EdgeHandling { get; set; }
 
-      /// <summary>
-      /// Velocity of the emitter in the X direction.
-      /// </summary>
+		/// <summary>
+		/// Velocity of the emitter in the X direction.
+		/// </summary>
 		Curve VelocityX { get; set; }
 
-      /// <summary>
-      /// Velocity of the emittter in the Y direction.
-      /// </summary>
+		/// <summary>
+		/// Velocity of the emittter in the Y direction.
+		/// </summary>
 		Curve VelocityY { get; set; }
 
-      /// <summary>
-      /// Manual position of the emitter in the X direction.
-      /// </summary>
+		/// <summary>
+		/// Manual position of the emitter in the X direction.
+		/// </summary>
 		Curve X { get; set; }
 
-      /// <summary>
-      /// Manual position of the emitter in the Y direction.
-      /// </summary>
+		/// <summary>
+		/// Manual position of the emitter in the Y direction.
+		/// </summary>
 		Curve Y { get; set; }
       
-      /// <summary>
-      /// Manual flow of the emitter.
-      /// </summary>
+		/// <summary>
+		/// Manual flow of the emitter.
+		/// </summary>
 		Curve Flow { get; set; }
 
-      /// <summary>
-      /// Source size of the emitter.
-      /// </summary>
+		/// <summary>
+		/// Source size of the emitter.
+		/// </summary>
 		Curve SourceSize { get; set; }
 
-      /// <summary>
-      /// Controls how the nozzle (angle) is positioned.
-      /// </summary>
+		/// <summary>
+		/// Controls how the nozzle (angle) is positioned.
+		/// </summary>
 		NozzleMovement NozzleMovement { get; set; }
 
-      /// <summary>
-      /// Angle of the emitter's nozzle.
-      /// </summary>
+		/// <summary>
+		/// Angle of the emitter's nozzle.
+		/// </summary>
 		Curve NozzleAngle { get; set; }
       	          		
-      /// <summary>
-      /// Start angle in degrees of the emitter oscillation.
-      /// </summary>
+		/// <summary>
+		/// Start angle in degrees of the emitter oscillation.
+		/// </summary>
 		int OscillateStartAngle { get; set; }
 
-      /// <summary>
-      /// End angle in degrees of the emitter oscillation.
-      /// </summary>
+		/// <summary>
+		/// End angle in degrees of the emitter oscillation.
+		/// </summary>
 		int OscillateEndAngle { get; set; }
 
-      /// <summary>
-      /// Speed of rotation of the emitter nozzle when rotating.
-      /// </summary>
+		/// <summary>
+		/// Speed of rotation of the emitter nozzle when rotating.
+		/// </summary>
 		Curve OscillationSpeed { get; set; }
 
-      /// <summary>
-      /// Determines how the flow is controlled for the emitter (Continuous, Pulsating, Use Marks).
-      /// </summary>
+		/// <summary>
+		/// Determines how the flow is controlled for the emitter (Continuous, Pulsating, Use Marks).
+		/// </summary>
 		FlowControl FlowControl { get; set; }
                  
-      /// <summary>
-      /// Name of the selected mark collection.
-      /// </summary>
+		/// <summary>
+		/// Name of the selected mark collection.
+		/// </summary>
 		string MarkCollectionName { get; set; }
 
-      /// <summary>
-      /// Guid of the selected mark collection.
-      /// </summary>
+		/// <summary>
+		/// Guid of the selected mark collection.
+		/// </summary>
 		Guid MarkCollectionId { get; set; }
 
-      /// <summary>
-      /// Collection of the mark collection names.
-      /// </summary>
+		/// <summary>
+		/// Collection of the mark collection names.
+		/// </summary>
 		ObservableCollection<string> MarkNameCollection { get; set; }
 
-      /// <summary>
-      /// Collection of the mark collections.
-      /// </summary>
+		/// <summary>
+		/// Collection of the mark collections.
+		/// </summary>
 		ObservableCollection<IMarkCollection> MarkCollections { get; set; }
 
 		// The following properties are not persisted but are used during rendering.
@@ -231,70 +219,70 @@ namespace VixenModules.Effect.Liquid
 		/// </summary>
 		int OnTime { get; set; }
 
-      /// <summary>
-      /// Off time for the emitter in seconds.
-      /// </summary>
+		/// <summary>
+		/// Off time for the emitter in seconds.
+		/// </summary>
 		int OffTime { get; set; }
 		
-      /// <summary>
-      /// Frame counter for the color array.
-      /// This counter determines when it is time to move to the next color in the array.
-      /// </summary>
+		/// <summary>
+		/// Frame counter for the color array.
+		/// This counter determines when it is time to move to the next color in the array.
+		/// </summary>
 		int FrameColorCounter { get; set; }
 
-      /// <summary>
-      /// Current index into the color array.
-      /// </summary>
+		/// <summary>
+		/// Current index into the color array.
+		/// </summary>
 		int ColorArrayIndex { get; set; }
 			
-      /// <summary>
-      /// Current X location of the emitter.
-      /// </summary>
+		/// <summary>
+		/// Current X location of the emitter.
+		/// </summary>
 		double LocationX { get; set; }
 
-      /// <summary>
-      /// Current Y location of the emitter.
-      /// </summary>
+		/// <summary>
+		/// Current Y location of the emitter.
+		/// </summary>
 		double LocationY { get; set; }
 
-      /// <summary>
-      /// Velocity of the emitter in the X axis.
-      /// </summary>
+		/// <summary>
+		/// Velocity of the emitter in the X axis.
+		/// </summary>
 		double VelX { get; set; }
 
-      /// <summary>
-      /// Velocity of the emitter in the Y axis.
-      /// </summary>
+		/// <summary>
+		/// Velocity of the emitter in the Y axis.
+		/// </summary>
 		double VelY { get; set; }
 
-      /// <summary>
-      /// Number frames the emitter has been On.
-      /// </summary>
+		/// <summary>
+		/// Number frames the emitter has been On.
+		/// </summary>
 		int OnTimer { get; set; }
 
-      /// <summary>
-      /// Number of frames the emitter has been Off.
-      /// </summary>
+		/// <summary>
+		/// Number of frames the emitter has been Off.
+		/// </summary>
 		int OffTimer { get; set; }		
 	
-      /// <summary>
-      /// True when the emitter is On.
-      /// </summary>
+		/// <summary>
+		/// True when the emitter is On.
+		/// </summary>
 		bool On { get; set; }
 
-      /// <summary>
-      /// Degrees to move when the emitter is rotating.
-      /// </summary>
+		/// <summary>
+		/// Degrees to move when the emitter is rotating.
+		/// </summary>
 		int AngleOscillationDelta { get; set; }
 
-      /// <summary>
-      /// True when the emitter is rotating clockwise.
-      /// </summary>
+		/// <summary>
+		/// True when the emitter is rotating clockwise.
+		/// </summary>
 		bool OscillationClockwise { get; set; }
 
-      /// <summary>
-      /// Angle of the emitter in integer precision.
-      /// </summary>
+		/// <summary>
+		/// Angle of the emitter in integer precision.
+		/// </summary>
 		int IntAngle { get; set; }
 	}
 }
